feat: updating and cleaning up script and making utils
