+++ scripts/gen-did/main.ts
@@ -1,97 +1,9 @@
 import * as Kilt from '@kiltprotocol/sdk-js'
-import * as fs from 'fs/promises'
 
-import {
-    blake2AsU8a,
-    keyExtractPath,
-    keyFromPath,
-    mnemonicGenerate,
-    mnemonicToMiniSecret,
-    sr25519PairFromSeed,
-} from '@polkadot/util-crypto'
-
-import { u8aToHex } from '@polkadot/util'
-
-interface DidSecrets {
-    did: string
-    authentication: {
-        pubKey: string
-        seed: string
-    }
-    attestation: {
-        pubKey: string
-        seed: string
-    }
-    keyAgreement: {
-        pubKey: string
-        seed: string
-        privKey: string
-    }
-}
-
-const signingKeyType = 'sr25519'
-
-function generateKeyAgreement(mnemonic: string): Kilt.KiltEncryptionKeypair {
-    const secretKeyPair = sr25519PairFromSeed(mnemonicToMiniSecret(mnemonic))
-    const { path } = keyExtractPath('//did//keyAgreement//0')
-    const { secretKey } = keyFromPath(secretKeyPair, path, signingKeyType)
-    return Kilt.Utils.Crypto.makeEncryptionKeypairFromSeed(blake2AsU8a(secretKey))
-}
-
-function generateKeypairs(mnemonic = mnemonicGenerate()): {
-    baseMnemonic: string
-    authentication: Kilt.KiltKeyringPair
-    keyAgreement: Kilt.KiltEncryptionKeypair
-    assertionMethod: Kilt.KiltKeyringPair
-} {
-    console.log('2', mnemonic)
-    const baseKey = Kilt.Utils.Crypto.makeKeypairFromSeed(mnemonicToMiniSecret(mnemonic), signingKeyType)
-
-    const authentication = baseKey.derive('//did//0') as typeof baseKey
-
-    const assertionMethod = baseKey.derive('//did//assertion//0') as typeof baseKey
-
-    const keyAgreement = generateKeyAgreement(mnemonic)
-
-    return {
-        baseMnemonic: mnemonic,
-        authentication: authentication,
-        keyAgreement: keyAgreement,
-        assertionMethod: assertionMethod,
-    }
-}
-
-function exportKeypairs(keypairs: {
-    baseMnemonic: string
-    authentication: Kilt.KiltKeyringPair
-    keyAgreement: Kilt.KiltEncryptionKeypair
-    assertionMethod: Kilt.KiltKeyringPair
-}): DidSecrets {
-    return {
-        did: `did:kilt:${keypairs.authentication.address}`,
-        authentication: {
-            pubKey: keypairs.authentication.address,
-            seed: keypairs.baseMnemonic + '//did//0',
-        },
-        attestation: {
-            pubKey: keypairs.assertionMethod.address,
-            seed: keypairs.baseMnemonic + '//did//assertion//0',
-        },
-        keyAgreement: {
-            pubKey: u8aToHex(keypairs.keyAgreement.publicKey),
-            seed: keypairs.baseMnemonic + '//did//keyAgreement//0',
-            privKey: u8aToHex(keypairs.keyAgreement.secretKey),
-        },
-    }
-}
-
-async function writeDidSecretsFile(didSecrets: DidSecrets, filename: string) {
-    await fs.writeFile(filename, JSON.stringify(didSecrets, null, 2))
-}
+import { checkAndWriteFile, exportKeypairs, generateKeypairs, signingKeyType } from './utils/utils'
 
 async function main() {
     // connect to kilt node
-
     let endpoint = 'wss://peregrine.kilt.io/'
     if (process.env.ENDPOINT) {
         switch (process.env.ENDPOINT) {
@@ -106,7 +18,6 @@ async function main() {
         }
     }
     console.debug(`Connecting to ${endpoint}...`)
-
     const api = await Kilt.connect(endpoint)
     console.debug(`Connected to ${endpoint}`)
     // console.log('api', api.on)
@@ -127,7 +38,7 @@ async function main() {
 
     const keypairs = generateKeypairs()
     const didSecrets = exportKeypairs(keypairs)
-    await writeDidSecretsFile(didSecrets, 'did-secrets.json')
+    await checkAndWriteFile(didSecrets, 'did-secrets.json')
     console.debug(`Wrote did-secrets.json`)
     const getStoreTxSignCallback: Kilt.Did.GetStoreTxSignCallback = async ({ data }) => ({
         signature: keypairs.authentication.sign(data),
@@ -159,7 +70,7 @@ async function main() {
         result = await Kilt.Did.resolve(`did:kilt:${keypairs.authentication.address}`)
     }
     // write the DID document to a file
-    await fs.writeFile('did-document.json', JSON.stringify(result.document, null, 2))
+    await checkAndWriteFile(result.document, 'did-document.json')
 
     await Kilt.disconnect()
 }
